Wizzer API V1 - Deprecated
Had to clear the commit history to avoid making sensitive data public.
All API keys, document numbers and phone numbers in this version were either removed or are random and invalid numbers.
With this in mind, for this version of the API to work properly you will have to replace those with valid values and keys.
Bugs caused by this removal are unknown, as I did not test the code after removing the values.